moved do_send_proposal back to send mod
